pyproject.toml
[project]
name = "mcp-auth-guard"
version = "1.0.0"
description = "Intuitive authorization middleware for MCP tools with type-safe policies"
authors = [
    { name = "MCP Auth Guard Contributors", email = "info@mcp-auth-guard.dev" }
]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.4",
    "pydantic>=2.11.7",
    "pyjwt[crypto]>=2.8.0",
    "pyyaml>=6.0.1",
    "typer>=0.15.2",
    "rich>=13.0.0",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.9.10",
    "mypy>=1.10.0",
]
examples = [
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
]

[project.scripts]
mcp-auth = "mcp_auth_guard.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_auth_guard"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
